combine db conections for blog and videos
